move dappier tools tests into it's own file
